+++ CHANGELOG.md
@@ -3,6 +3,8 @@
     - `taxonkit lca`:
         - add a new flag `-b/--buffer-size` to set the size of the line buffer. [#75](https://github.com/shenwei356/taxonkit/issues/75)
         - fix typo: `--separater` -> `--separater`, the former is still available for backward compatibility.
+    - `taxonkit reformat`:
+        - output compatible format for TaxIds not found in the database. [#79](https://github.com/shenwei356/taxonkit/issues/79)
 - [TaxonKit v0.14.1](https://github.com/shenwei356/taxonkit/releases/tag/v0.14.1)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.14.1/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.14.1)
     - `taxonkit reformat`: